grpc-examples/python/0_buf/pyproject.toml
[project]
name = "0-buf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
]
